requirements.txt
opencv-contrib-python
numpy
matplotlib
tensorflow
keras
python-socketio
Flask
eventlet
streamlit==1.20.0
torch==1.10.2
torchvision==0.11.3
grad-cam==1.4.6
scikit-learn==1.0.2